% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDTRE_download_functions.R
\name{download_municipality}
\alias{download_municipality}
\title{Function to download BDTRE database for each municipality desired.
You can use the get_municipality_codes function to select municipality codes by province or find just the municipality you are interested on}
\usage{
download_municipality(municipality_number)
}
\arguments{
\item{municipality_number}{It should be a data frame structured as the one created by the get_municipality_codes function or download_selected_mun function, inputting
a series of municipality codes to be downloaded}
}
\value{
It does not return anything inside the code, it is just a download action and it organizes in different folders what has been downloaded
}
\description{
Function to download BDTRE database for each municipality desired.
You can use the get_municipality_codes function to select municipality codes by province or find just the municipality you are interested on
}
\examples{
user_list <- c('003016', '005008', '006004', '002035')
download_selected_mun(user_list)
user_list <- download_selected_mun(user_list)
download_municipality(user_list)
}
