% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDTRE_download_functions.R
\name{download_selected_mun}
\alias{download_selected_mun}
\title{It helps the user to set a list to input inside the download_municipality function in order to download the desired municipalities}
\usage{
download_selected_mun(user_list)
}
\arguments{
\item{user_list}{It should be a vector built buy the user with the municipal codes on which get information}
}
\value{
It returns a data frame similar to the one made by get_municipality_codes function, but just for the selected codes by the user
}
\description{
It helps the user to set a list to input inside the download_municipality function in order to download the desired municipalities
}
\examples{
user_list <- c('003016', '005008', '006004', '002035')
download_selected_mun(user_list)
user_list <- download_selected_mun(user_list)
}
