% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDTRE_download_functions.R
\name{get_municipality_codes}
\alias{get_municipality_codes}
\title{Function to get municipality identification codes}
\usage{
get_municipality_codes(selected_province)
}
\arguments{
\item{selected_province}{The variable to input to get the municipality codes. The user can select the municipalities by province putting as input '001' (Torino), '002' (Vercelli), '003' (Novara), '004' (Cuneo), '005' (Asti), '006' (Alessandria), '096' (Biella), '103' (Verbano) or 'All' to save the municipality codes for all the region}
}
\value{
It returns a table with column names divided by: Code province, Code municipality, Municipality name
}
\description{
Function to get municipality identification codes
}
\examples{
get_municipality_codes('001')
}
