% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDTRE_download_functions.R
\name{merging_shp}
\alias{merging_shp}
\title{This function is useful for merging all the shape files for the selected category.
It helps the user to make an unique shape file from all the shape files related to one category}
\usage{
merging_shp(chosen_shp)
}
\arguments{
\item{chosen_shp}{Must be any from selection_list}
}
\value{
nc_merge
}
\description{
This function is useful for merging all the shape files for the selected category.
It helps the user to make an unique shape file from all the shape files related to one category
}
\examples{
nc_merge <- merging_shp('limi_comuni_piem')

}
