% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDTRE_download_functions.R
\name{percent_urbanized}
\alias{percent_urbanized}
\title{Example of possible uses of the package and downloaded data
Calculate urbanization index for each municipalities
To do that it keeps all the shape files for the urbanized areas calculating the areas
for each of it and summarizing and after make the percent on the total of the municipality surface}
\usage{
percent_urbanized(m_codes)
}
\arguments{
\item{m_codes}{Must be a data frame such as from get_municipality_codes}
}
\value{
data_frame_data You can return a data frame where the data are stored in different columns:
1) name of the municipalities, 2) ISTAT code and 3) percent of urbanized areas
}
\description{
Example of possible uses of the package and downloaded data
Calculate urbanization index for each municipalities
To do that it keeps all the shape files for the urbanized areas calculating the areas
for each of it and summarizing and after make the percent on the total of the municipality surface
}
\examples{
Biella_province <- get_municipality_codes("096")
download_municipality(Biella_province)
data_frame_data <- percent_urbanized(Biella_province)
}
