% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BDTRE_download_functions.R
\name{selection_path}
\alias{selection_path}
\title{Function for connecting each shape file inside the downloaded folders.
This function is the core of the package and is used to read each file downloaded using the download_municipality function.
In order to be read, the shape file must in fact be inserted in a path}
\usage{
selection_path(chosen_shp)
}
\arguments{
\item{chosen_shp}{The user must enter in this field a selection as a string among the possible ones which are:
'lim_com', 'limi_comuni_piem', 'sed_amm', 'ab_cda_vert', 'el_idr_vert', 'nd_idr', 'af_acq',
'dre_sup', 'ghi_nv', 'invaded_vert', 'sp_acq', 'aatt', 'attr_sp', 'cr_edf', 'cs_edi',
'edi_min', 'ele_cp', 'un_vol', 'embankment', 'dam', 'edifc', 'galler', 'man_tr', 'mn_arr',
'mn_int', 'mn_ind', 'mn_mau', 'mu_sos', 'bridge', 'tralic', 'acc_int'
'acc_pc_civico_tp_str', 'es_amm', 'tp_str', 'cv_liv_class', 'f_nter', 'pt_quo', 'a_tras',
'riverbed', 'alveo_a', 'pe_uins', 'cv_aes', 'cv_dis', 'loc_sg', 'scr_cr',
'ar_vrd', 'bosco', 'cl_agr', 'for_pc', 'ps_inc', 'a_pveg', 'tree', 'fil_al', 'for_pc',
'ac_vei', 'ar_vms', 'el_str_tp_str', 'el_vms', 'gz_str', 'gz_vms', 'iz_str', 'tr_str',
'ac_cic', 'ac_ped', 'el_fer', 'gz_fer', 'sd_fer', 'nd_ele', 'tr_ele', 'v_rete'.

It is possible to view the jupyter notebook file for the meaning of each.}
}
\value{
path_shp
}
\description{
Function for connecting each shape file inside the downloaded folders.
This function is the core of the package and is used to read each file downloaded using the download_municipality function.
In order to be read, the shape file must in fact be inserted in a path
}
\examples{
paths <- selection_path('limi_comuni_piem')
# or you can simple see the paths
selection_path('limi_comuni_piem')
}
